chore: get wallet working
