+++ .env.example
@@ -1,1 +1,4 @@
-WEB3AUTH_CLIENT_ID=BEglQSgt4cUWcj6SKRdu5QkOXTsePmMcusG5EAoyjyOYKlVRjIF1iCNnMOTfpzCiunHRrMui8TIwQPXdkQ8Yxuk
+NEXT_PUBLIC_WEB3AUTH_CLIENT_ID=BEglQSgt4cUWcj6SKRdu5QkOXTsePmMcusG5EAoyjyOYKlVRjIF1iCNnMOTfpzCiunHRrMui8TIwQPXdkQ8Yxuk
+NEXT_PUBLIC_WEB3AUTH_CLIENT_SECRET=
+NEXT_PUBLIC_RPC_KEY=
+NEXT_PUBLIC_RPC_URL=
