+++ package.json
@@ -11,11 +11,12 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.0.18",
-    "@web3auth/base": "^7.1.0",
-    "@web3auth/metamask-adapter": "^7.1.0",
-    "@web3auth/modal": "^7.1.1",
-    "@web3auth/torus-evm-adapter": "^7.1.0",
-    "@web3auth/torus-wallet-connector-plugin": "^7.1.1",
+    "@web3auth/base": "^7.0.4",
+    "@web3auth/metamask-adapter": "^7.0.4",
+    "@web3auth/modal": "^7.0.5",
+    "@web3auth/torus-evm-adapter": "^7.0.4",
+    "@web3auth/torus-wallet-connector-plugin": "^7.0.4",
+    "@web3auth/wallet-connect-v2-adapter": "^7.0.4",
     "ethers": "^6.8.1",
     "next": "^14.0.1",
     "react": "^18",
