+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-import Home from "@/components/Home";
+import Dashboard from "@/components/Dashboard";
 
 const HomePage = () => {
   return (
     <main>
-      <Home />
+      <Dashboard />
     </main>
   );
 };
