+++ src/components/Dashboard.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-export default function Home() {
+export default function Dashboard() {
   return (
     <div className="">
       <main className="isolate">
