+++ src/components/Navbar.tsx
@@ -1,15 +1,14 @@
 "use client";
 
-import { Fragment } from "react";
+import { Fragment, useEffect, useState } from "react";
 import { Disclosure, Menu, Transition } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
-import Image from "next/image";
+import RPC from "../utils/ethersRPC";
 import Link from "next/link";
-// import logo from "../app/assets/logo.svg";
-
-const user = {
-  address: "0x0",
-};
+import { IProvider, CHAIN_NAMESPACES, WALLET_ADAPTERS } from "@web3auth/base";
+import { Web3Auth } from "@web3auth/modal";
+import { TorusWalletConnectorPlugin } from "@web3auth/torus-wallet-connector-plugin";
+import { OpenloginAdapter } from "@web3auth/openlogin-adapter";
 
 const navigation = [
   { name: "Dashboard", href: "/", current: true },
@@ -22,11 +21,371 @@ const userNavigation = [
   { name: "Sign out", href: "/log-out" },
 ];
 
+const clientId =
+  process.env.NEXT_PUBLIC_WEB3AUTH_CLIENT_ID!;
+const rpcUrl = process.env.NEXT_PUBLIC_RPC_URL!;
+
 function classNames(...classes: string[]) {
   return classes.filter(Boolean).join(" ");
 }
 
 export default function Navbar() {
+  const [web3auth, setWeb3auth] = useState<Web3Auth | null>(null);
+  const [torusPlugin, setTorusPlugin] =
+    useState<TorusWalletConnectorPlugin | null>(null);
+  const [provider, setProvider] = useState<IProvider | null>(null);
+  const [loggedIn, setLoggedIn] = useState(false);
+
+  useEffect(() => {
+    const init = async () => {
+      try {
+        const web3auth = new Web3Auth({
+          clientId,
+          chainConfig: {
+            chainNamespace: CHAIN_NAMESPACES.EIP155,
+            chainId: "5",
+            rpcTarget: rpcUrl,
+          },
+          web3AuthNetwork: "sapphire_devnet",
+        });
+
+        const openloginAdapter = new OpenloginAdapter({
+          loginSettings: {
+            mfaLevel: "optional",
+          },
+          adapterSettings: {
+            whiteLabel: {
+              appName: "IncentiVibe",
+              logoLight: "https://web3auth.io/images/w3a-L-Favicon-1.svg",
+              logoDark: "https://web3auth.io/images/w3a-D-Favicon-1.svg",
+              defaultLanguage: "en", // en, de, ja, ko, zh, es, fr, pt, nl
+              mode: "dark", // whether to enable dark mode. defaultValue: false
+            },
+            mfaSettings: {
+              deviceShareFactor: {
+                enable: true,
+                priority: 1,
+                mandatory: true,
+              },
+              backUpShareFactor: {
+                enable: true,
+                priority: 2,
+                mandatory: false,
+              },
+              socialBackupFactor: {
+                enable: true,
+                priority: 3,
+                mandatory: false,
+              },
+              passwordFactor: {
+                enable: true,
+                priority: 4,
+                mandatory: false,
+              },
+            },
+            loginConfig: {
+              google: {
+                verifier: "w3a-google-demo",
+                typeOfLogin: "google",
+                clientId:
+                  "519228911939-cri01h55lsjbsia1k7ll6qpalrus75ps.apps.googleusercontent.com", //use your app client id you got from google
+              },
+              facebook: {
+                verifier: "w3a-facebook-demo",
+                typeOfLogin: "facebook",
+                clientId: "215892741216994", //use your app client id you got from facebook
+              },
+              discord: {
+                verifier: "w3a-discord-demo",
+                typeOfLogin: "discord",
+                clientId: "1151006428610433095", //use your app client id you got from discord
+              },
+              twitch: {
+                verifier: "w3a-twitch-demo",
+                typeOfLogin: "twitch",
+                clientId: "3k7e70gowvxjaxg71hjnc8h8ih3bpf", //use your app client id you got from twitch
+              },
+            },
+          },
+        });
+        web3auth.configureAdapter(openloginAdapter);
+
+        // plugins and adapters are optional and can be added as per your requirement
+        // read more about plugins here: https://web3auth.io/docs/sdk/web/plugins/
+
+        // adding torus wallet connector plugin
+
+        const torusPlugin = new TorusWalletConnectorPlugin({
+          torusWalletOpts: {},
+          walletInitOptions: {
+            whiteLabel: {
+              theme: { isDark: true, colors: { primary: "#00a8ff" } },
+              logoDark: "https://web3auth.io/images/w3a-L-Favicon-1.svg",
+              logoLight: "https://web3auth.io/images/w3a-D-Favicon-1.svg",
+            },
+            useWalletConnect: true,
+          },
+        });
+        setTorusPlugin(torusPlugin);
+        await web3auth.addPlugin(torusPlugin);
+
+        setWeb3auth(web3auth);
+
+        // await web3auth.initModal();
+
+        await web3auth.initModal({
+          modalConfig: {
+            [WALLET_ADAPTERS.OPENLOGIN]: {
+              label: "openlogin",
+              loginMethods: {
+                // Disable facebook and reddit
+                apple: {
+                  name: "facebook",
+                  showOnModal: true,
+                },
+                reddit: {
+                  name: "reddit",
+                  showOnModal: true,
+                },
+                line: {
+                  name: "line",
+                  showOnModal: false,
+                },
+                github: {
+                  name: "github",
+                  showOnModal: true,
+                },
+                wechat: {
+                  name: "wechat",
+                  showOnModal: false,
+                },
+                twitter: {
+                  name: "twitter",
+                  showOnModal: true,
+                },
+                kakao: {
+                  name: "kakao",
+                  showOnModal: false,
+                },
+                linkedin: {
+                  name: "linkedin",
+                  showOnModal: true,
+                },
+                weibo: {
+                  name: "weibo",
+                  showOnModal: false,
+                },
+                // Disable email_passwordless and sms_passwordless
+                email_passwordless: {
+                  name: "email_passwordless",
+                  showOnModal: true,
+                },
+                sms_passwordless: {
+                  name: "sms_passwordless",
+                  showOnModal: true,
+                },
+              },
+            },
+            [WALLET_ADAPTERS.METAMASK]: {
+              label: "metamask",
+              loginMethods: {
+                metamask: {
+                  name: "metamask",
+                  showOnModal: true,
+                },
+              },
+            },
+          },
+        });
+
+        if (web3auth.provider) {
+          setProvider(web3auth.provider);
+        }
+
+        if (web3auth.connected) {
+          setLoggedIn(true);
+        }
+      } catch (error) {
+        console.error(error);
+      }
+    };
+
+    init();
+  }, []);
+
+  const login = async () => {
+    uiConsole("Logging in...");
+    alert("Logging in...");
+    if (!web3auth) {
+      uiConsole("web3auth not initialized yet");
+
+      return false;
+    }
+    const web3authProvider = await web3auth.connect();
+    setProvider(web3authProvider);
+    setLoggedIn(true);
+
+    return true;
+  };
+
+  const authenticateUser = async () => {
+    if (!web3auth) {
+      uiConsole("web3auth not initialized yet");
+
+      return;
+    }
+    const idToken = await web3auth.authenticateUser();
+    uiConsole(idToken);
+  };
+
+  const getUserInfo = async () => {
+    if (!web3auth) {
+      uiConsole("web3auth not initialized yet");
+
+      return;
+    }
+    const user = await web3auth.getUserInfo();
+    uiConsole(user);
+  };
+
+  const logout = async () => {
+    if (!web3auth) {
+      uiConsole("web3auth not initialized yet");
+
+      return;
+    }
+    await web3auth.logout();
+    setProvider(null);
+    setLoggedIn(false);
+  };
+
+  const showWCM = async () => {
+    if (!torusPlugin) {
+      uiConsole("torus plugin not initialized yet");
+
+      return;
+    }
+    torusPlugin.showWalletConnectScanner();
+    uiConsole();
+  };
+
+  const initiateTopUp = async () => {
+    if (!torusPlugin) {
+      uiConsole("torus plugin not initialized yet");
+
+      return;
+    }
+    torusPlugin.initiateTopup("moonpay", {
+      selectedAddress: "0x8cFa648eBfD5736127BbaBd1d3cAe221B45AB9AF",
+      selectedCurrency: "USD",
+      fiatValue: 100,
+      selectedCryptoCurrency: "ETH",
+      chainNetwork: "mainnet",
+    });
+  };
+
+  const getChainId = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const rpc = new RPC(provider);
+    const chainId = await rpc.getChainId();
+    uiConsole(chainId);
+  };
+
+  const addChain = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const newChain = {
+      chainId: "0x5",
+      displayName: "Goerli",
+      chainNamespace: CHAIN_NAMESPACES.EIP155,
+      tickerName: "Goerli",
+      ticker: "ETH",
+      decimals: 18,
+      rpcTarget: "https://rpc.ankr.com/eth_goerli",
+      blockExplorer: "https://goerli.etherscan.io",
+    };
+    await web3auth?.addChain(newChain);
+    uiConsole("New Chain Added");
+  };
+
+  const switchChain = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    await web3auth?.switchChain({ chainId: "0x5" });
+    uiConsole("Chain Switched");
+  };
+
+  const getAccounts = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const rpc = new RPC(provider);
+    const address = await rpc.getAccounts();
+    uiConsole(address);
+  };
+
+  const getBalance = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+      return;
+    }
+    const rpc = new RPC(provider);
+    const balance = await rpc.getBalance();
+    uiConsole(balance);
+  };
+
+  const sendTransaction = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const rpc = new RPC(provider);
+    const receipt = await rpc.sendTransaction();
+    uiConsole(receipt);
+  };
+
+  const signMessage = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const rpc = new RPC(provider);
+    const signedMessage = await rpc.signMessage();
+    uiConsole(signedMessage);
+  };
+
+  const getPrivateKey = async () => {
+    if (!provider) {
+      uiConsole("provider not initialized yet");
+
+      return;
+    }
+    const rpc = new RPC(provider);
+    const privateKey = await rpc.getPrivateKey();
+    uiConsole(privateKey);
+  };
+
+  function uiConsole(...args: any[]): void {
+    const el = document.querySelector("#console>p");
+    if (el) {
+      el.innerHTML = JSON.stringify(args || {}, null, 2);
+    }
+  }
+
   return (
     <Disclosure
       as="nav"
@@ -79,6 +438,14 @@ export default function Navbar() {
                   <button
                     type="button"
                     className="relative inline-flex border border-white text-white hover:bg-white hover:text-black items-center gap-x-1.5 rounded-md px-3 py-2 text-sm font-semibold"
+                    onClick={async () => {
+                      console.log("login");
+
+                      // log the user in
+                      await login();
+
+                      console.log("login status:", loggedIn);
+                    }}
                   >
                     {/* Add wallet connect button */}
                     Connect
