+++ src/context/Providers.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import { Web3AuthContextProvider } from "./Web3AuthContext";
-
 const Providers = ({ children }: { children: React.ReactNode }) => {
-  return <Web3AuthContextProvider>{children}</Web3AuthContextProvider>;
+  return <div>{children}</div>;
 };
 
 export default Providers;
